Support for reading ScVal contract parameters and fuzzing over them (#17)
* Add a soroban contract that uses the various SCVals

* Create SCType class that can read the bindings json from soroban CLI

* Start implementing errors. Raise an error when reading an unsupported
SCType.

* SCVals that store the actual values, and hypothesis strategies to
generate them

* to_kast for SCValues

* Fuzz over the test contracts

* Add a failing test

* More explicit naming of test contracts

* Test the various supported scvaltypes

* Set Version: 0.1.12

* pyupgrade

* Set Version: 0.1.13

---------

+++ src/tests/integration/data/soroban/contracts/test_adder/src/lib.rs
@@ -6,13 +6,13 @@ pub struct AdderContract;
 
 #[contractimpl]
 impl AdderContract {
-    pub fn add(env: Env, first: u32, second: u32) -> u32 {
-       first + second
+    pub fn add(env: Env, first: u32, second: u32) -> u64 {
+       first as u64 + second as u64
     }
 
     pub fn test_add(env: Env, num: u32) -> bool {
         let sum = Self::add(env, num, 5);
-        sum == num + 5
+        sum == num as u64 + 5
     }
 }
